Move models to firebase directory
