+++ src/app/firebase/services/firebase-room.service.ts
@@ -3,14 +3,14 @@ import { AngularFirestore } from '@angular/fire/compat/firestore';
 import { fromPromise } from 'rxjs/internal/observable/innerFrom';
 import { Observable } from 'rxjs';
 import firebaseConstants from '../firebaseConstants';
-import { RoomCreationRequest } from '../../room-creation/models/room-creation-request';
+import { Room } from '../models/room';
 
 @Injectable()
 export class FirebaseRoomService {
 
   constructor(private firestore: AngularFirestore) { }
 
-  add(room: RoomCreationRequest): Observable<string> {
+  add(room: Room): Observable<string> {
     return fromPromise(this.firestore.collection(firebaseConstants.collections.rooms)
       .add(room)
       .then(roomReference => roomReference.id));
