+++ src/app/store/actions.ts
@@ -1,5 +1,5 @@
 ﻿import { createAction, props } from '@ngrx/store';
-import { Player } from '../player-creation/models/player';
+import { Player } from '../firebase/models/player';
 
 export const setPlayer = createAction('[Room Joining] Set Player', props<{ player: Player }>());
 export const removePlayer = createAction('[Room Leaving] Remove Player');
